cleanup some protocol stuff

+++ lib/bus/iec/protocol/cbmstandardserial.h
@@ -33,7 +33,7 @@ namespace Protocol
     class IecProtocolSerial : public IECProtocol
     {
         private:
-        int16_t receiveBits();
+        int8_t receiveBits();
         bool sendBits(uint8_t data);
 
 
@@ -52,7 +52,7 @@ namespace Protocol
          * @brief receive byte from bus
          * @return The byte received from bus.
          */
-        virtual int16_t receiveByte();
+        virtual int8_t receiveByte();
         
         /**
          * @brief send byte to bus
